Add basic supervisor functionality
The Supervisor trait lets an actor know when a worker
beneath it has crashed. (TODO: change name), by sending it
the id of the crash and some panic info. (TODO: return id
from spawn function).

This is a rather cool PoC which I want to extend to also
handle process calls which returns an Result::Err.